refactor: abstract enum `Step` and `UploadHandler` class in a separate file
