+++ apps/examples/nextjs/src/types/step.ts
@@ -1,0 +1,7 @@
+export enum Step {
+    CHOOSE_FILE,
+    REQUEST_UPLOAD_URL,
+    UPLOAD_FILE,
+    CONFIRM_UPLOAD,
+    POLL_FILE,
+}
